feat(image): adds method to create Associations from remote sources
