+++ pkg/image/association.go
@@ -5,25 +5,10 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/fs"
-	"io/ioutil"
-	"os"
 	"path/filepath"
 
-	ctrsimgmanifest "github.com/containers/image/v5/manifest"
-	imgspecv1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/openshift/oc/pkg/cli/image/imagesource"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
 )
-
-type ErrInvalidComponent struct {
-	image string
-	tag   string
-}
-
-func (e *ErrInvalidComponent) Error() string {
-	return fmt.Sprintf("image %q has invalid component %q", e.image, e.tag)
-}
 
 // Associations is a map for Association
 // searching
@@ -37,7 +22,7 @@ type AssociationSet map[string]Associations
 type Association struct {
 	// Name of the image.
 	Name string `json:"name"`
-	// Path to image data within archive.
+	// Path to image in new location (archive or registry)
 	Path string `json:"path"`
 	// ID of the image. Joining this value with "manifests" and Path
 	// will produce a path to the image's manifest.
@@ -144,9 +129,8 @@ func (as AssociationSet) ContainsKey(setKey, key string) (found bool) {
 
 // Merge Associations into the receiver.
 func (as AssociationSet) Merge(in AssociationSet) {
-	for _, imageName := range in.Keys() {
-		values, _ := in.Search(imageName)
-		for _, value := range values {
+	for imageName, assocs := range in {
+		for _, value := range assocs {
 			as.Add(imageName, value)
 		}
 	}
@@ -172,8 +156,7 @@ func (as *AssociationSet) Decode(r io.Reader) error {
 		return fmt.Errorf("error decoding image associations: %v", err)
 	}
 	// Update paths for local usage.
-	for _, imageName := range as.Keys() {
-		assocs, _ := as.Search(imageName)
+	for imageName, assocs := range *as {
 		for _, assoc := range assocs {
 			assoc.Path = filepath.FromSlash(assoc.Path)
 			if err := as.UpdateValue(imageName, assoc); err != nil {
@@ -182,6 +165,34 @@ func (as *AssociationSet) Decode(r io.Reader) error {
 		}
 	}
 	return nil
+}
+
+// UpdatePath path will update path values for local
+// AssociationSet use
+func (as *AssociationSet) UpdatePath() error {
+	// Update paths for local usage.
+	for imageName, assocs := range *as {
+		for _, assoc := range assocs {
+			assoc.Path = filepath.FromSlash(assoc.Path)
+			if err := as.UpdateValue(imageName, assoc); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+// GetDigests will return all layer and manifest digests in the association set
+func (as *AssociationSet) GetDigests() []string {
+	var digests []string
+	for _, assocs := range *as {
+		for _, assoc := range assocs {
+			digests = append(digests, assoc.LayerDigests...)
+			digests = append(digests, assoc.ManifestDigests...)
+			digests = append(digests, assoc.ID)
+		}
+	}
+	return digests
 }
 
 func (as AssociationSet) validate() error {
@@ -219,146 +230,16 @@ func (as AssociationSet) validate() error {
 	return utilerrors.NewAggregate(errs)
 }
 
-func AssociateImageLayers(rootDir string, imgMappings TypedImageMapping) (AssociationSet, utilerrors.Aggregate) {
-	errs := []error{}
-	bundleAssociations := AssociationSet{}
+func GetImageFromBlob(as AssociationSet, digest string) string {
+	for imageName, assocs := range as {
+		for _, assoc := range assocs {
+			for _, dgst := range assoc.LayerDigests {
+				if dgst == digest {
+					return imageName
+				}
 
-	skipParse := func(ref string) bool {
-		seen := bundleAssociations.SetContainsKey(ref)
-		return seen
-	}
-
-	localRoot := filepath.Join(rootDir, "v2")
-	for image, diskLoc := range imgMappings {
-		if diskLoc.Type != imagesource.DestinationFile {
-			errs = append(errs, fmt.Errorf("image destination for %q is not type file", image.Ref.Exact()))
-			continue
-		}
-		dirRef := diskLoc.Ref.AsRepository().String()
-		imagePath := filepath.Join(localRoot, dirRef)
-
-		// Verify that the dirRef exists before proceeding
-		if _, err := os.Stat(imagePath); err != nil {
-			errs = append(errs, fmt.Errorf("image %q mapping %q: %v", image, dirRef, err))
-			continue
-		}
-
-		var tagOrID string
-		if diskLoc.Ref.Tag != "" {
-			tagOrID = diskLoc.Ref.Tag
-		} else {
-			tagOrID = diskLoc.Ref.ID
-		}
-
-		if tagOrID == "" {
-			errs = append(errs, &ErrInvalidComponent{image.String(), tagOrID})
-			continue
-		}
-
-		// TODO(estroz): parallelize
-		associations, err := associateImageLayers(image.Ref.String(), localRoot, dirRef, tagOrID, "oc-mirror", image.Category, skipParse)
-		if err != nil {
-			errs = append(errs, err)
-			continue
-		}
-		for _, association := range associations {
-			bundleAssociations.Add(image.Ref.String(), association)
-		}
-	}
-
-	return bundleAssociations, utilerrors.NewAggregate(errs)
-}
-
-func associateImageLayers(image, localRoot, dirRef, tagOrID, defaultTag string, typ ImageType, skipParse func(string) bool) (associations []Association, err error) {
-	if skipParse(image) {
-		return nil, nil
-	}
-
-	manifestPath := filepath.Join(localRoot, filepath.FromSlash(dirRef), "manifests", tagOrID)
-	// TODO(estroz): this file mode checking block is likely only necessary
-	// for the first recursion leaf since image manifest layers always contain id's,
-	// so unroll this component into AssociateImageLayers.
-
-	info, err := os.Lstat(manifestPath)
-	if errors.Is(err, os.ErrNotExist) {
-		return nil, &ErrInvalidComponent{image, tagOrID}
-	} else if err != nil {
-		return nil, err
-	}
-	// Tags are always symlinks due to how `oc` libraries mirror manifest files.
-	id, tag := tagOrID, tagOrID
-	switch m := info.Mode(); {
-	case m&fs.ModeSymlink != 0:
-		// Tag is the file name, so follow the symlink to the layer ID-named file.
-		dst, err := os.Readlink(manifestPath)
-		if err != nil {
-			return nil, fmt.Errorf("error evaluating image tag symlink: %v", err)
-		}
-		id = filepath.Base(dst)
-	case m.IsRegular():
-		// Layer ID is the file name, and no tag exists.
-		tag = defaultTag
-		if defaultTag != "" {
-			// If set, add a subset of the digest to randomize the
-			// tag in the event multiple digests are pulled for the same
-			// image
-			tag = defaultTag + id[7:13]
-			manifestDir := filepath.Dir(manifestPath)
-			symlink := filepath.Join(manifestDir, tag)
-			if err := os.Symlink(info.Name(), symlink); err != nil {
-				return nil, err
 			}
 		}
-	default:
-		return nil, fmt.Errorf("expected symlink or regular file mode, got: %b", m)
 	}
-	manifestBytes, err := ioutil.ReadFile(filepath.Clean(manifestPath))
-	if err != nil {
-		return nil, fmt.Errorf("error reading image manifest file: %v", err)
-	}
-
-	association := Association{
-		Name:       image,
-		Path:       dirRef,
-		ID:         id,
-		TagSymlink: tag,
-		Type:       typ,
-	}
-	switch mt := ctrsimgmanifest.GuessMIMEType(manifestBytes); mt {
-	case "":
-		return nil, errors.New("unparseable manifest mediaType")
-	case imgspecv1.MediaTypeImageIndex, ctrsimgmanifest.DockerV2ListMediaType:
-		list, err := ctrsimgmanifest.ListFromBlob(manifestBytes, mt)
-		if err != nil {
-			return nil, err
-		}
-		for _, instance := range list.Instances() {
-			digestStr := instance.String()
-			// Add manifest references so publish can recursively look up image layers
-			// for the manifests of this list.
-			association.ManifestDigests = append(association.ManifestDigests, digestStr)
-			// Recurse on child manifests, which should be in the same directory
-			// with the same file name as it's digest.
-			childAssocs, err := associateImageLayers(digestStr, localRoot, dirRef, digestStr, "", typ, skipParse)
-			if err != nil {
-				return nil, err
-			}
-			associations = append(associations, childAssocs...)
-		}
-	default:
-		// Treat all others as image manifests.
-		manifest, err := ctrsimgmanifest.FromBlob(manifestBytes, mt)
-		if err != nil {
-			return nil, err
-		}
-		for _, layerInfo := range manifest.LayerInfos() {
-			association.LayerDigests = append(association.LayerDigests, layerInfo.Digest.String())
-		}
-		// The config is just another blob, so associate it opaquely.
-		association.LayerDigests = append(association.LayerDigests, manifest.ConfigInfo().Digest.String())
-	}
-
-	associations = append(associations, association)
-
-	return associations, nil
+	return ""
 }
