+++ pkg/image/credentials.go
@@ -1,4 +1,4 @@
-package config
+package image
 
 import (
 	"errors"
